rename library folders to avoid name collisions
