--- projects/unified-angular-directory/package.json
@@ -1,12 +1,0 @@
-{
-  "name": "unified-angular-directory",
-  "version": "0.0.1",
-  "peerDependencies": {
-    "@angular/common": "^16.1.0",
-    "@angular/core": "^16.1.0"
-  },
-  "dependencies": {
-    "tslib": "^2.3.0"
-  },
-  "sideEffects": false
-}
