--- projects/unified-angular-directory/src/lib/unified-angular-directory.module.ts
@@ -1,9 +1,0 @@
-import { NgModule } from '@angular/core';
-import { UnifiedAngularDirectoryComponent } from './unified-angular-directory.component';
-
-@NgModule({
-    declarations: [],
-    imports: [UnifiedAngularDirectoryComponent],
-    exports: [UnifiedAngularDirectoryComponent],
-})
-export class UnifiedAngularDirectoryModule {}
