--- projects/unified-angular-directory/src/public-api.ts
@@ -1,6 +1,0 @@
-/*
- * Public API Surface of unified-angular-directory
- */
-
-export * from './lib/unified-angular-directory.component';
-export * from './lib/unified-angular-directory.module';
