+++ projects/unified-directory/ng-package.json
@@ -1,6 +1,6 @@
 {
   "$schema": "../../node_modules/ng-packagr/ng-package.schema.json",
-  "dest": "../../dist/unified-angular-directory",
+  "dest": "../../dist/unified-directory",
   "lib": {
     "entryFile": "src/public-api.ts"
   }
