+++ projects/unified-directory/src/lib/pipes/filter-categories.pipe.ts
