+++ projects/unified-directory/src/lib/unified-directory.module.ts
@@ -1,0 +1,9 @@
+import { NgModule } from '@angular/core';
+import { UnifiedDirectory } from './unified-directory.component';
+
+@NgModule({
+  declarations: [],
+  imports: [UnifiedDirectory],
+  exports: [UnifiedDirectory],
+})
+export class UnifiedAngularDirectoryModule {}
