+++ projects/unified-directory/src/public-api.ts
@@ -1,0 +1,5 @@
+/*
+ * Public API Surface of @unified-api/angular-directory
+ */
+
+export { UnifiedDirectory } from './lib/unified-directory.component';
